+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+* 2.0.0
+ - Renames `getInstance` to `findInstance`.
+
+* 1.1.1
+  - Functional components doesn't return any instance now
+
 * 1.1.0
   - Working for nested and higher order component
   - Added changelog